fix run notebook tests
